+++ tests/testthat/setup-run-model.R
@@ -59,7 +59,8 @@ a_hintr_calibration_options <- list(
   spectrum_aware_calibration_level = "national",
   spectrum_aware_calibration_strat = "age_coarse",
   spectrum_infections_calibration_level = "none",
-  spectrum_infections_calibration_strat = "age_coarse"
+  spectrum_infections_calibration_strat = "age_coarse",
+  calibrate_method = "logistic"
 )
 
 ## Use fit.RDS if it exists locally, otherwise just use the actual functions
@@ -137,22 +138,6 @@ MockSimpleProgress <- R6::R6Class(
     }
   )
 )
-
-clone_output <- function(output) {
-  output_path <- tempfile()
-  file.copy(output$output_path, output_path)
-  spectrum_path <- tempfile(fileext = ".zip")
-  file.copy(output$spectrum_path, spectrum_path)
-  coarse_output_path <- tempfile(fileext = ".zip")
-  file.copy(output$coarse_output_path, coarse_output_path)
-  summary_report_path <- tempfile(fileext = ".html")
-  file.copy(output$summary_report_path, summary_report_path)
-  calibration_path <- tempfile(fileext = ".rds")
-  file.copy(output$calibration_path, calibration_path)
-  build_hintr_output(output_path, spectrum_path, coarse_output_path,
-                     summary_report_path, calibration_path,
-                     output$metadata)
-}
 
 expect_file_equivalent <- function(path_object, path_expected) {
   object_md5 <- tools::md5sum(path_object)