+++ tests/testthat/test-run-model.R
@@ -289,6 +289,7 @@ test_that("setting rng_seed returns same output", {
   options$spectrum_artnum_calibration_level <- "none"
   options$spectrum_aware_calibration_level <- "none"
   options$spectrum_infections_calibration_level <- "none"
+  options$calibrate_method <- "logistic"
 
   output_path <- tempfile()
   output_spectrum <- tempfile(fileext = ".zip")
@@ -422,13 +423,46 @@ test_that("input data types can be formatted", {
 })
 
 test_that("model run can be calibrated", {
-  ## Calibration modifies files in place. We want to make sure we don't modify
-  ## the test setup output and introduce race condition in tests.
-  ## Create a copy and calibrate that
-  output <- clone_output(a_hintr_output)
-  calibrated_output <- hintr_calibrate(output, a_hintr_calibration_options)
+
+  ## Calibration makes no modification of existing files.
+  output_hash <- tools::md5sum(a_hintr_output$output_path)
+  spectrum_hash <- tools::md5sum(a_hintr_output$spectrum_path)
+  coarse_output_hash <- tools::md5sum(a_hintr_output$coarse_output_path)
+  summary_report_hash <- tools::md5sum(a_hintr_output$summary_report_path)
+  calibration_hash <- tools::md5sum(a_hintr_output$calibration_path)
+
+  output_path <- tempfile()
+  spectrum_path <- tempfile(fileext = ".zip")
+  coarse_output_path <- tempfile(fileext = ".zip")
+  summary_report_path = tempfile(fileext = ".html")
+  calibration_path <- tempfile(fileext = ".rds")
+  calibrated_output <- hintr_calibrate(a_hintr_output,
+                                       a_hintr_calibration_options,
+                                       output_path,
+                                       spectrum_path,
+                                       coarse_output_path,
+                                       summary_report_path,
+                                       calibration_path)
 
   expect_s3_class(calibrated_output, "hintr_output")
+  expect_equal(calibrated_output$output_path, output_path)
+  expect_equal(calibrated_output$spectrum_path, spectrum_path)
+  expect_equal(calibrated_output$coarse_output_path, coarse_output_path)
+  expect_equal(calibrated_output$summary_report_path, summary_report_path)
+  expect_equal(calibrated_output$calibration_path, calibration_path)
+  expect_equal(calibrated_output$metadata, a_hintr_output$metadata)
+
+  ## Calibration does not modify original files
+  expect_equal(tools::md5sum(a_hintr_output$output_path),
+               output_hash)
+  expect_equal(tools::md5sum(a_hintr_output$spectrum_path),
+               spectrum_hash)
+  expect_equal(tools::md5sum(a_hintr_output$coarse_output_path),
+               coarse_output_hash)
+  expect_equal(tools::md5sum(a_hintr_output$summary_report_path),
+               summary_report_hash)
+  expect_equal(tools::md5sum(a_hintr_output$calibration_path),
+               calibration_hash)
 
   ## Output has been calibrated
   expect_file_different(calibrated_output$output_path,
@@ -437,7 +471,7 @@ test_that("model run can be calibrated", {
   ## Check there is some data
   expect_equal(nrow(indicators_output), 31 * 3 * 3 * 22 * 12 + 3 * 9 * 22 * 11)
 
-  ## Spectrum file has been calibrated
+  ## Spectrum file has been calibrated & original files unchanged
   expect_file_different(calibrated_output$spectrum_path,
                         a_hintr_output$spectrum_path)
   file_list <- unzip(calibrated_output$spectrum_path, list = TRUE)
@@ -502,9 +536,6 @@ test_that("model run can be calibrated", {
   expect_equal(calibrated_output$metadata, a_hintr_output$metadata)
 
   ## Can calibrate multiple times
-  ## Copy files again as calibration modified in place and we want to test
-  ## that recalibrating returns us new results
-  clone <- clone_output(calibrated_output)
   calibration_options <- list(
     spectrum_plhiv_calibration_level = "national",
     spectrum_plhiv_calibration_strat = "sex_age_coarse",
@@ -513,9 +544,11 @@ test_that("model run can be calibrated", {
     spectrum_aware_calibration_level = "subnational",
     spectrum_aware_calibration_strat = "age_coarse",
     spectrum_infections_calibration_level = "none",
-    spectrum_infections_calibration_strat = "age_coarse"
+    spectrum_infections_calibration_strat = "age_coarse",
+    calibrate_method = "logistic"
   )
-  calibrated_output_2 <- hintr_calibrate(clone, calibration_options)
+  calibrated_output_2 <- hintr_calibrate(calibrated_output,
+                                         calibration_options)
 
   expect_s3_class(calibrated_output_2, "hintr_output")
 
@@ -815,10 +848,9 @@ test_that("simple progress", {
 
 test_that("calibration reports simple progress", {
   mock_new_simple_progress <- mockery::mock(MockSimpleProgress$new())
-  output <- clone_output(a_hintr_output)
   with_mock("naomi:::new_simple_progress" = mock_new_simple_progress, {
     messages <- naomi_evaluate_promise(
-      hintr_calibrate(output, a_hintr_calibration_options))
+      hintr_calibrate(a_hintr_output, a_hintr_calibration_options))
   })
   expect_length(messages$progress, 3)
   progress <- messages$progress
@@ -838,7 +870,8 @@ test_that("validate_calibrate_options errors if required options are missing", {
     "spectrum_artnum_calibration_level" = "none",
     "spectrum_artnum_calibration_strat" = "none",
     "spectrum_aware_calibration_level" = "none",
-    "spectrum_aware_calibration_strat" = "none")),
+    "spectrum_aware_calibration_strat" = "none",
+    "calibrate_method" = "logistic")),
     paste0("Calibration cannot be run, missing options for ",
            "spectrum_infections_calibration_level, ",
            "spectrum_infections_calibration_strat."))
@@ -852,7 +885,8 @@ test_that("validate_calibrate_options errors if required options are missing", {
     "spectrum_artnum_calibration_strat" = "none",
     "spectrum_aware_calibration_level" = "none",
     "spectrum_aware_calibration_strat" = "none",
-    "spectrum_infections_calibration_level" = "none")),
+    "spectrum_infections_calibration_level" = "none",
+    "calibrate_method" = "logistic")),
     paste0("Calibration cannot be run, missing options for ",
            "spectrum_infections_calibration_strat."))
 
@@ -865,5 +899,33 @@ test_that("validate_calibrate_options errors if required options are missing", {
     "spectrum_aware_calibration_level" = "none",
     "spectrum_aware_calibration_strat" = "none",
     "spectrum_infections_calibration_level" = "none",
-    "spectrum_infections_calibration_strat" = "none")))
+    "spectrum_infections_calibration_strat" = "none",
+    "calibrate_method" = "logistic")))
+
+  expect_error(validate_calibrate_options(list(
+    "spectrum_plhiv_calibration_level" = "none",
+    "spectrum_plhiv_calibration_level" = "none",
+    "spectrum_plhiv_calibration_strat" = "none",
+    "spectrum_artnum_calibration_level" = "none",
+    "spectrum_artnum_calibration_strat" = "none",
+    "spectrum_aware_calibration_level" = "none",
+    "spectrum_aware_calibration_strat" = "none",
+    "spectrum_infections_calibration_level" = "none",
+    "spectrum_infections_calibration_strat" = "none")),
+    paste0("Calibration cannot be run, missing options for ",
+           "calibrate_method."))
+
+  expect_error(validate_calibrate_options(list(
+    "spectrum_plhiv_calibration_level" = "none",
+    "spectrum_plhiv_calibration_level" = "none",
+    "spectrum_plhiv_calibration_strat" = "none",
+    "spectrum_artnum_calibration_level" = "none",
+    "spectrum_artnum_calibration_strat" = "none",
+    "spectrum_aware_calibration_level" = "none",
+    "spectrum_aware_calibration_strat" = "none",
+    "spectrum_infections_calibration_level" = "none",
+    "spectrum_infections_calibration_strat" = "none",
+    "calibrate_method" = "JIBBERISH")),
+    paste0("calibrate_method must be either \"logistic\" or \"proportional\""))
+
 })