+++ NEWS.md
@@ -1,7 +1,28 @@
+# naomi 2.3.1
+
+* Set default for `calibrate_method` model option in user interface via JSON metadata.
+* Fix swapped indicator_code and indicator_id for Data Pack metadata which were swapped.
+* Refactor [`calibrate_outputs()`]; handle uncertainty ranges for proportion adjustments.
+
+# naomi 2.3.0
+
+* Implement 'logistic' scaling option for [`calibrate_outputs()`] such that estimates are
+  adjusted on logistic scale by fine district/sex/age group to ensure proportions do 
+  not go above 100%.
+  - Note: implementation does not yet handle uncertainty ranges. Those might still go above
+    100%.
+* Fix to model calibration for number aware of status to align with proportion aware 
+  of status and number unaware of status.
+
+
+# naomi 2.2.5
+
+* Add BWA, HTI, and COD to PEPFAR Data Pack PSNU list
+* Fix Datim data element UID for indicator DIAGNOSED_SUBNAT.T_1.
+
 # naomi 2.2.4
 
-* Replace `readLines()` with `brio::read_lines()`.
-* Use `readr::read_csv()` as default file reader in function `naomi_read_csv()`.
+* hintr_calibrate will store calibration outputs to new files instead of overwriting existing
 
 # naomi 2.2.3